Begin refactoring game login into game.lua
